requirements.txt
# Django Core y API REST
Django>=5.0,<5.3
djangorestframework>=3.14,<3.17
djangorestframework-simplejwt>=5.0,<5.6
django-cors-headers>=4.0,<4.8

# Base de datos (mantenemos SQLite por simplicidad)
# psycopg2-binary>=2.9,<2.10  # Comentado: Eliminamos PostgreSQL/PostGIS

# Reconocimiento y síntesis de voz
vosk>=0.3.45
gtts>=2.4.0
pydub>=0.25.1

# Procesamiento de lenguaje natural
scikit-learn>=1.3.0
nltk>=3.8.1

# Peticiones HTTP para APIs externas
requests>=2.31.0

# Utilidades para audio
SpeechRecognition>=3.10.0
pyaudio>=0.2.11  # Para captura de audio en tiempo real

# Procesamiento geoespacial (sin PostGIS)
geopy>=2.4.0

# Utilidades adicionales
python-decouple>=3.8  # Para gestión de variables de entorno
whitenoise>=6.5.0  # Para servir archivos estáticos en producción 